Unique sessionId verifcation for incoming web socket connections
After the external client passes its pubKeyHex to the server, a unique session Id is created to establish a web-socket connection for the client's pubKey.
The server waits for an incoming web-socket conneciton, verifies that the sessionIdf has been signed with the pubKeys signature, keeps the connection opens if the signature is valid, or closes it if invalid.

Removed the timeout functions for the core web-socket client and identity files.

Signed-off-by: brioux <Bertrand.rioux@gmail.com>